新增 ubuntu/debian 基础命令脚本

+++ AIO/Scripts/basic_command.sh
@@ -34,9 +34,9 @@ white(){
 command_choose() {
     clear
     echo "=================================================================="
-    echo -e "\t\tLinux基础命令脚本 by 忧郁滴飞叶"
+    echo -e "\t\tubuntu/debian 基础命令脚本 by 忧郁滴飞叶"
     echo -e "\t\n"  
-    echo "欢迎使用Linux基础命令脚本"
+    echo "欢迎使用 ubuntu/debian 基础命令脚本"
     echo "请选择要执行的服务："
     echo "=================================================================="
     echo "1. 启动服务（程序）"
@@ -69,29 +69,28 @@ command_choose() {
         5)
             systemctl daemon-reload    
             green "重新加载配置完成"
-            rm -rf /mnt/basic_command.sh    #delete   
+            [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete    
             ;;
         6)
             all_sotp_restart
-            rm -rf /mnt/basic_command.sh    #delete   
             ;;
         7)
             check_error 
             ;;            
         8)    
             clear
-            rm -rf /mnt/basic_command.sh    #delete   
+            [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete    
             ;;            
         9)
             quick
             ;;            
         0)
             red "退出脚本，感谢使用."
-            rm -rf /mnt/basic_command.sh    #delete             
+            [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete              
             ;;
         -)
             white "脚本切换中，请等待..."
-            rm -rf /mnt/basic_command.sh    #delete       
+            [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete        
             wget -q -O /mnt/main_install.sh https://raw.githubusercontent.com/feiye2021/LinuxScripts/main/AIO/Scripts/main_install.sh && chmod +x /mnt/main_install.sh && /mnt/main_install.sh
             ;;
         *)
@@ -113,34 +112,34 @@ sotp_restart() {
     elif [[ "$action_name" == "restart" ]]; then
         green "重启 ${service_name} 成功"
     fi
-    rm -rf /mnt/basic_command.sh    #delete   
+    [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete    
 }
 ################################ 一键停止、重加载、重启服务 ################################
 all_sotp_restart() {
     read -p "请输入执行操作的服务（程序）的名称： " service_name
     systemctl stop ${service_name} && systemctl daemon-reload && systemctl restart ${service_name}
     green "一键停止、重加载、重启 ${service_name} 服务成功"
-    rm -rf /mnt/basic_command.sh    #delete   
+    [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete    
 }
 ################################ 查看服务（程序）报错日志 ################################
 check_error() {
     read -p "请输入要查询的服务（程序）的名称： " service_name
-    rm -rf /mnt/basic_command.sh    #delete
+    [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete 
     green "开始输出错误日志："
     journalctl -xe | grep ${service_name}
 }
 ################################ 转快速启动 ################################
 quick() {
     echo "=================================================================="
-    echo -e "\t Linux基础命令脚本转快速启动 by 忧郁滴飞叶"
+    echo -e "\t ubuntu/debian 基础命令脚本转快速启动 by 忧郁滴飞叶"
     echo -e "\t\n"  
-    echo -e "欢迎使用Linux基础命令脚本转快速启动脚本，脚本运行完成后\n在shell界面输入command即可调用脚本"
+    echo -e "欢迎使用 ubuntu/debian 基础命令脚本转快速启动脚本，脚本\n运行完成后在shell界面输入 esay 即可调用脚本"
     echo "=================================================================="
     white "开始转快速启动..."
-    wget -O /usr/bin/command https://raw.githubusercontent.com/feiye2021/LinuxScripts/main/AIO/Scripts/basic_command.sh
-    chmod +x /usr/bin/command
-    green "Linux基础命令脚本转快捷启动已完成，shell界面输入command即可调用脚本"
-    rm -rf /mnt/basic_command.sh    #delete   
+    wget -O /usr/bin/easy https://raw.githubusercontent.com/feiye2021/LinuxScripts/main/AIO/Scripts/basic_command.sh
+    chmod +x /usr/bin/easy
+    green "ubuntu/debian 基础命令脚本转快捷启动已完成，shell界面输入 easy 即可调用脚本"
+    [ -f /mnt/basic_command.sh ] && rm -rf /mnt/basic_command.sh    #delete    
 }
 ################################ 主程序 ################################
 command_choose
